+++ packages/react-sdk/src/components/CallParticipantsList/CallParticipantListingItem.tsx
@@ -10,6 +10,7 @@ import {
   Restricted,
   useCall,
   useConnectedUser,
+  useI18n,
 } from '@stream-io/video-react-bindings';
 import {
   OwnCapability,
@@ -42,14 +43,16 @@ export const CallParticipantListingItem = ({
   const isVideoOn = participant.publishedTracks.includes(
     SfuModels.TrackType.VIDEO,
   );
-  const isPinned = !!participant.pinnedAt;
+  const isPinned = !!participant.pin;
+
+  const { t } = useI18n();
 
   return (
     <div className="str-video__participant-listing-item">
       <DisplayName participant={participant} />
       <div className="str-video__participant-listing-item__media-indicator-group">
         <MediaIndicator
-          title={isAudioOn ? 'Microphone on' : 'Microphone off'}
+          title={isAudioOn ? t('Microphone on') : t('Microphone off')}
           className={clsx(
             'str-video__participant-listing-item__icon',
             `str-video__participant-listing-item__icon-${
@@ -58,7 +61,7 @@ export const CallParticipantListingItem = ({
           )}
         />
         <MediaIndicator
-          title={isVideoOn ? 'Camera on' : 'Camera off'}
+          title={isVideoOn ? t('Camera on') : t('Camera off')}
           className={clsx(
             'str-video__participant-listing-item__icon',
             `str-video__participant-listing-item__icon-${
@@ -68,7 +71,7 @@ export const CallParticipantListingItem = ({
         />
         {isPinned && (
           <MediaIndicator
-            title={'Pinned'}
+            title={t('Pinned')}
             className={clsx(
               'str-video__participant-listing-item__icon',
               'str-video__participant-listing-item__icon-pinned',
@@ -95,12 +98,13 @@ type DisplayNameProps = {
 // todo: implement display device flag
 const DefaultDisplayName = ({ participant }: DisplayNameProps) => {
   const connectedUser = useConnectedUser();
+  const { t } = useI18n();
 
-  const meFlag = participant.userId === connectedUser?.id ? 'Me' : '';
-  const nameOrId = participant.name || participant.userId || 'Unknown';
+  const meFlag = participant.userId === connectedUser?.id ? t('Me') : '';
+  const nameOrId = participant.name || participant.userId || t('Unknown');
   let displayName;
   if (!participant.name) {
-    displayName = meFlag || nameOrId || 'Unknown';
+    displayName = meFlag || nameOrId || t('Unknown');
   } else if (meFlag) {
     displayName = `${nameOrId} (${meFlag})`;
   } else {
@@ -139,49 +143,71 @@ export const ParticipantActionsContextMenu = ({
     document.pictureInPictureElement,
   );
   const activeCall = useCall();
+  const { t } = useI18n();
+
+  const { pin, publishedTracks, sessionId, userId } = participant;
+
+  const hasAudio = publishedTracks.includes(SfuModels.TrackType.AUDIO);
+  const hasVideo = publishedTracks.includes(SfuModels.TrackType.VIDEO);
+  const hasScreenShare = publishedTracks.includes(
+    SfuModels.TrackType.SCREEN_SHARE,
+  );
 
   const blockUser = () => {
-    activeCall?.blockUser(participant.userId);
+    activeCall?.blockUser(userId);
   };
-
-  // FIXME: soft kicking does not work this way
-  // also needs to be session-based
-  // const kickUserClickHandler = () => {
-  //   getCall()?.updateCallMembers({
-  //     remove_members: [participant.userId],
-  //     disconnectRemovedMembers: true,
-  //   });
-  // };
-
   const muteAudio = () => {
-    activeCall?.muteUser(participant.userId, 'audio');
+    activeCall?.muteUser(userId, 'audio');
   };
   const muteVideo = () => {
-    activeCall?.muteUser(participant.userId, 'video');
+    activeCall?.muteUser(userId, 'video');
   };
   const muteScreenShare = () => {
-    activeCall?.muteUser(participant.userId, 'screenshare');
+    activeCall?.muteUser(userId, 'screenshare');
   };
 
   const grantPermission = (permission: string) => () => {
     activeCall?.updateUserPermissions({
-      user_id: participant.userId,
+      user_id: userId,
       grant_permissions: [permission],
     });
   };
 
   const revokePermission = (permission: string) => () => {
     activeCall?.updateUserPermissions({
-      user_id: participant.userId,
+      user_id: userId,
       revoke_permissions: [permission],
     });
   };
 
   const toggleParticipantPinnedAt = () => {
-    activeCall?.setParticipantPinnedAt(
-      participant.sessionId,
-      participant.pinnedAt ? undefined : Date.now(),
-    );
+    if (pin) {
+      activeCall?.unpin(sessionId);
+    } else {
+      activeCall?.pin(sessionId);
+    }
+  };
+
+  const pinForEveryone = () => {
+    activeCall
+      ?.pinForEveryone({
+        user_id: userId,
+        session_id: sessionId,
+      })
+      .catch((err) => {
+        console.error(`Failed to pin participant ${userId}`, err);
+      });
+  };
+
+  const unpinForEveryone = () => {
+    activeCall
+      ?.unpinForEveryone({
+        user_id: userId,
+        session_id: sessionId,
+      })
+      .catch((err) => {
+        console.error(`Failed to unpin participant ${userId}`, err);
+      });
   };
 
   const toggleFullscreenMode = () => {
@@ -236,92 +262,100 @@ export const ParticipantActionsContextMenu = ({
 
   return (
     <GenericMenu>
-      <GenericMenuButtonItem onClick={toggleParticipantPinnedAt}>
+      <GenericMenuButtonItem
+        onClick={toggleParticipantPinnedAt}
+        disabled={pin && !pin.isLocalPin}
+      >
         <Icon icon="pin" />
-        {participant.pinnedAt ? 'Unpin' : 'Pin'}
+        {pin ? t('Unpin') : t('Pin')}
       </GenericMenuButtonItem>
+      <Restricted requiredGrants={[OwnCapability.PIN_FOR_EVERYONE]}>
+        <GenericMenuButtonItem
+          onClick={pinForEveryone}
+          disabled={pin && !pin.isLocalPin}
+        >
+          <Icon icon="pin" />
+          {t('Pin for everyone')}
+        </GenericMenuButtonItem>
+        <GenericMenuButtonItem
+          onClick={unpinForEveryone}
+          disabled={!pin || pin.isLocalPin}
+        >
+          <Icon icon="pin" />
+          {t('Unpin for everyone')}
+        </GenericMenuButtonItem>
+      </Restricted>
       <Restricted requiredGrants={[OwnCapability.BLOCK_USERS]}>
         <GenericMenuButtonItem onClick={blockUser}>
           <Icon icon="not-allowed" />
-          Block
+          {t('Block')}
         </GenericMenuButtonItem>
       </Restricted>
-      {/* <GenericMenuButtonItem disabled onClick={kickUserClickHandler}>
-        Kick
-      </GenericMenuButtonItem> */}
       <Restricted requiredGrants={[OwnCapability.MUTE_USERS]}>
-        <GenericMenuButtonItem
-          disabled={
-            !participant.publishedTracks.includes(SfuModels.TrackType.VIDEO)
-          }
-          onClick={muteVideo}
-        >
+        <GenericMenuButtonItem disabled={!hasVideo} onClick={muteVideo}>
           <Icon icon="camera-off-outline" />
-          Turn off video
+          {t('Turn off video')}
         </GenericMenuButtonItem>
         <GenericMenuButtonItem
-          disabled={
-            !participant.publishedTracks.includes(
-              SfuModels.TrackType.SCREEN_SHARE,
-            )
-          }
+          disabled={!hasScreenShare}
           onClick={muteScreenShare}
         >
           <Icon icon="screen-share-off" />
-          Turn off screen share
+          {t('Turn off screen share')}
         </GenericMenuButtonItem>
-        <GenericMenuButtonItem
-          disabled={
-            !participant.publishedTracks.includes(SfuModels.TrackType.AUDIO)
-          }
-          onClick={muteAudio}
-        >
+        <GenericMenuButtonItem disabled={!hasAudio} onClick={muteAudio}>
           <Icon icon="no-audio" />
-          Mute audio
+          {t('Mute audio')}
         </GenericMenuButtonItem>
       </Restricted>
       {participantViewElement && (
         <GenericMenuButtonItem onClick={toggleFullscreenMode}>
-          {fullscreenModeOn ? 'Leave' : 'Enter'} fullscreen
+          {t('{{ direction }} fullscreen', {
+            direction: fullscreenModeOn ? t('Leave') : t('Enter'),
+          })}
         </GenericMenuButtonItem>
       )}
       {videoElement && document.pictureInPictureEnabled && (
         <GenericMenuButtonItem onClick={togglePictureInPicture}>
-          {pictureInPictureElement === videoElement ? 'Leave' : 'Enter'}{' '}
-          picture-in-picture
+          {t('{{ direction }} picture-in-picture', {
+            direction:
+              pictureInPictureElement === videoElement
+                ? t('Leave')
+                : t('Enter'),
+          })}
         </GenericMenuButtonItem>
       )}
       <Restricted requiredGrants={[OwnCapability.UPDATE_CALL_PERMISSIONS]}>
         <GenericMenuButtonItem
           onClick={grantPermission(OwnCapability.SEND_AUDIO)}
         >
-          Allow audio
+          {t('Allow audio')}
         </GenericMenuButtonItem>
         <GenericMenuButtonItem
           onClick={grantPermission(OwnCapability.SEND_VIDEO)}
         >
-          Allow video
+          {t('Allow video')}
         </GenericMenuButtonItem>
         <GenericMenuButtonItem
           onClick={grantPermission(OwnCapability.SCREENSHARE)}
         >
-          Allow screen sharing
+          {t('Allow screen sharing')}
         </GenericMenuButtonItem>
 
         <GenericMenuButtonItem
           onClick={revokePermission(OwnCapability.SEND_AUDIO)}
         >
-          Disable audio
+          {t('Disable audio')}
         </GenericMenuButtonItem>
         <GenericMenuButtonItem
           onClick={revokePermission(OwnCapability.SEND_VIDEO)}
         >
-          Disable video
+          {t('Disable video')}
         </GenericMenuButtonItem>
         <GenericMenuButtonItem
           onClick={revokePermission(OwnCapability.SCREENSHARE)}
         >
-          Disable screen sharing
+          {t('Disable screen sharing')}
         </GenericMenuButtonItem>
       </Restricted>
     </GenericMenu>