+++ sample-apps/react-native/dogfood/src/constants/A11yLabels.ts
@@ -1,4 +1,0 @@
-export enum A11yButtons {
-  LOG_OUT = 'log-out-button',
-  CHOOSE_MODE = 'choose-mode-button',
-}